feat: :art: Extra - CentroPolicies
Políticas creadas

+++ app/Http/Controllers/CentroController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreRequest;
 use App\Models\Centro;
+use App\Policies\CentroPolicy;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Http\Request;
 
@@ -28,6 +29,7 @@ class CentroController extends Controller
      */
     public function create(Centro $centro)
     {
+        $this->authorize('create',$centro);
         return view('centros.create', compact('centro'));
     }
 
@@ -53,6 +55,7 @@ class CentroController extends Controller
      */
     public function show(Centro $centro)
     {
+        $this->authorize('view',$centro);
         return view('centros.show',compact('centro'));
     }
 
@@ -64,6 +67,7 @@ class CentroController extends Controller
      */
     public function edit(Centro $centro)
     {
+        $this->authorize('edit',$centro);
         return view('centros.edit',compact('centro'));
     }
 
@@ -76,6 +80,7 @@ class CentroController extends Controller
      */
     public function update(Request $request, Centro $centro)
     {
+        $this->authorize('update',$centro);
         $centro->update($request->all());
         return redirect()->route('centros.index', $centro);
     }
@@ -88,6 +93,7 @@ class CentroController extends Controller
      */
     public function destroy(Centro $centro)
     {
+        $this->authorize('delete',$centro);
         $centro->delete();
         return redirect()->route('centros.index');
     }
